Move showDeps to corresponding CLI command
