+++ cli/commands/deps.js
@@ -1,4 +1,4 @@
-const { showDependencies } = require("../../internals/deps");
+const packageConfig = require("../../package/packageConfig");
 
 exports.command = "deps";
 exports.desc = "Manage clio dependencies";
@@ -11,3 +11,23 @@ exports.builder = yargs => {
 exports.handler = () => {
   showDependencies();
 };
+
+/**
+ * @method showDependencies
+ * @returns {void}
+ * @description Prints to the stdout the list of
+ *              dependencies listed in package.json
+ */
+
+function showDependencies() {
+  if (!packageConfig.hasClioDependencies()) {
+    console.log("No dependencies found in package.json");
+    return;
+  }
+
+  const deps = packageConfig.getPackageDependencies();
+  const formattedDeps = deps
+    .map(dep => `~> ${dep["name"]}: ${dep["version"]}`)
+    .join("\n");
+  console.log(formattedDeps);
+}
